+++ CHANGES.rst
@@ -14,6 +14,7 @@ in progress
 - Update dependencies: Use Jinja2 3.x; Remove configparser, it is built into Python 3
 - Add support for Python 3.11
 - Pushover service plugin: Enable passing of parameters ``html``, ``url``, ``url_title``
+- Add test for exercising configuration file scaffolding
 
 
 2021-10-31 0.28.1